+++ sdm-api-doorbell.groovy
@@ -10,12 +10,12 @@
  *  from the copyright holder
  *  Software is provided without warranty and your use of it is at your own risk.
  *
- *  version: 0.1.1
+ *  version: 0.2.0
  */
 
 metadata {
     definition(name: 'Google Nest Doorbell', namespace: 'dkilgore90', author: 'David Kilgore', importUrl: 'https://raw.githubusercontent.com/dkilgore90/google-sdm-api/master/sdm-api-doorbell.groovy') {
-       // capability 'VideoCamera'
+        //capability 'VideoCamera'
         capability 'PushableButton'
         capability 'ImageCapture'
         capability 'Refresh'
@@ -33,9 +33,14 @@ metadata {
     }
     
     preferences {
-        input 'minimumMotionTime', 'number', title: 'Motion timeout', 'description': 'minimum time (in seconds) that the motion attribute will show `active` after receiving an event', required: true, defaultValue: 15
-        input 'minimumPresenseTime', 'number', title: 'Presense timeout', 'description': 'minimum time (in seconds) that the presence attribute will show `present` after receiving an event', required: true, defaultValue: 15
-         input 'minimumSoundTime', 'number', title: 'Sound timeout', 'description': 'minimum time (in seconds) that the sound attribute will show `detected` after receiving an event', required: true, defaultValue: 15
+        input 'minimumMotionTime', 'number', title: 'Motion timeout', description: 'minimum time (in seconds) that the motion attribute will show `active` after receiving an event', required: true, defaultValue: 15
+        input 'minimumPresenceTime', 'number', title: 'Presence timeout', description: 'minimum time (in seconds) that the presence attribute will show `present` after receiving an event', required: true, defaultValue: 15
+        input 'minimumSoundTime', 'number', title: 'Sound timeout', description: 'minimum time (in seconds) that the sound attribute will show `detected` after receiving an event', required: true, defaultValue: 15
+    
+        input 'chimeImageCapture', 'bool', title: 'Chime - Capture image?', description: 'whether to download the still image for "Chime" events', required: true, defaultValue: true
+        input 'personImageCapture', 'bool', title: 'Person - Capture image?', description: 'whether to download the still image for "Person" events', required: true, defaultValue: true
+        input 'motionImageCapture', 'bool', title: 'Motion - Capture image?', description: 'whether to download the still image for "Motion" events', required: true, defaultValue: true
+        input 'soundImageCapture', 'bool', title: 'Sound - Capture image?', description: 'whether to download the still image for "Sound" events', required: true, defaultValue: true
     }
 }
 
@@ -62,10 +67,10 @@ def refresh() {
 
 def processPerson() {
     device.sendEvent(name: 'presence', value: 'present')
-    if (minimumPresenseTime == null) {
-        device.updateSetting('minimumPresenseTime', 15)
+    if (minimumPresenceTime == null) {
+        device.updateSetting('minimumPresenceTime', 15)
     }
-    runIn(minimumPresenseTime, presenceInactive, [overwrite: true])
+    runIn(minimumPresenceTime, presenceInactive, [overwrite: true])
 }
 
 def presenceInactive() {
@@ -94,6 +99,26 @@ def processSound() {
 
 def soundInactive() {
     device.sendEvent(name: 'sound', value: 'not detected')
+}
+
+def shouldGetImage(String event) {
+    switch (event) {
+    case 'Chime':
+        return chimeImageCapture != null ? chimeImageCapture : true
+        break
+    case 'Person':
+        return personImageCapture != null ? personImageCapture : true
+        break
+    case 'Motion':
+        return motionImageCapture != null ? motionImageCapture : true
+        break
+    case 'Sound':
+        return soundImageCapture != null ? soundImageCapture : true
+        break
+    default:
+        return true
+        break
+    }
 }
 
 def take() {